+++ packages/expense_repository/lib/src/services/firebase_income_repo.dart
@@ -1,4 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:expense_repository/src/services/firebase_wallet_repo.dart';
+import 'package:expense_repository/src/models/wallet.dart';
 import '../models/income.dart';
 
 class FireStoreIncomeService{
@@ -40,9 +42,48 @@ class FireStoreIncomeService{
   }
 
   // delete
-  Future<void> deleteIncome(String docID){
-    return incomeCollection.doc(docID).delete();
+  // Future<void> deleteIncome(String docID){
+  //   return incomeCollection.doc(docID).delete();
+  // }
+  Future<void> deleteIncome(String docID) async {
+    try {
+      // Fetch the income document
+      DocumentSnapshot incomeDoc = await incomeCollection.doc(docID).get();
+      if (!incomeDoc.exists) {
+        throw Exception("Income not found");
+      }
+
+      // Convert document to Income object
+      Income incomeToDelete = Income.fromJson(incomeDoc.data() as Map<String, dynamic>);
+
+      // Fetch the wallet document ID
+      FireStoreWalletService walletService = FireStoreWalletService();
+      String walletDocId = await walletService.getWalletDocIdByUserId(incomeToDelete.userId);
+
+      // Fetch the current wallet using the fetched doc ID
+      Wallet currentWallet = await walletService.walletCollection.doc(walletDocId).get()
+          .then((doc) => Wallet.fromJson(doc.data() as Map<String, dynamic>));
+
+      // Calculate the updated balance by subtracting the income amount from the current balance
+      int updatedBalance = currentWallet.currentBalance - incomeToDelete.amount;
+
+      // Update wallet's current balance
+      Wallet updatedWallet = currentWallet.copy(
+        currentBalance: updatedBalance,
+        lastModified: DateTime.now(),
+      );
+
+      // Save the updated wallet back to Firestore
+      await walletService.updateWallet(walletDocId, updatedWallet);
+
+      // Delete the income document
+      await incomeCollection.doc(docID).delete();
+    } catch (e) {
+      // Handle errors appropriately
+      print('Failed to delete income: $e');
+    }
   }
+
 
   Stream<QuerySnapshot> getIncomeFromDateRange(String startDate, String endDate, String userId) {
     final incomeInDateRange = incomeCollection.
